Separate out the new functions.
